requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-RESTX==1.1.0
PyMySQL==1.1.0
marshmallow==3.19.0
pytest==7.3.1
python-dotenv==1.0.0
cryptography==40.0.2 